+++ Servicos/UseCases/UserUseCases/IDeleteUserRequest.cs
@@ -1,10 +1,12 @@
-﻿using FluentValidation;
+﻿using FluentResults;
+using FluentValidation;
 using MediatR;
+using Servicos.Erros;
 using Servicos.RepositoryInterfaces;
 
 namespace Servicos.UseCases.UserUseCases
 {
-    public class DeleteUserRequest : IRequest
+    public class DeleteUserRequest : IRequest<Result>
     {
         public int Id { get; set; }
 
@@ -17,20 +19,25 @@ namespace Servicos.UseCases.UserUseCases
         }
     }
 
-    public class DeleteUserRequestHandler : IRequestHandler<DeleteUserRequest>
+    public class DeleteUserRequestHandler : IRequestHandler<DeleteUserRequest, Result>
     {
         private readonly IUserRepository userRepository;
         public DeleteUserRequestHandler(IUserRepository userRepository)
         {
             this.userRepository = userRepository;
         }
-        public async Task Handle(DeleteUserRequest request, CancellationToken cancellationToken)
+        public async Task<Result> Handle(DeleteUserRequest request, CancellationToken cancellationToken)
         {
             var usuario = await userRepository.FiltrarUsuarioPorId(request.Id);
 
+            if (usuario is null)
+                return new NaoEncontrado("Usuario nao encontrado");
+
             userRepository.DeleteUsuario(usuario);
 
-            userRepository.Savechanges();
+            await userRepository.Savechanges();
+
+            return Result.Ok();
         }
     }
 }
